typecast to uint64_t for ros::Time
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@6455 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ highlevel/executive_trex/executive_trex_pr2/test/groundtruthtransform.cpp
@@ -13,7 +13,7 @@ public:
   std_msgs::Point32 msg;
   tf::TransformListener tf;
 
-  GroundTruthTransform() : ros::node("GroundTruthTransform"), tf(*this, true, 10000000000ULL) {
+  GroundTruthTransform() : ros::node("GroundTruthTransform"), tf(*this, true, (uint64_t)10000000000ULL) {
     advertise<std_msgs::Point32>("groundtruthposition");
   }
 
@@ -21,7 +21,7 @@ public:
     tf::Stamped<tf::Pose> robotPose, globalPose;
     robotPose.setIdentity();
     robotPose.frame_id_ = "base";
-    robotPose.stamp_ = ros::Time(0ULL);
+    robotPose.stamp_ = ros::Time((uint64_t)0ULL);
     try {
       this->tf.transformPose("map", robotPose, globalPose);
     } catch(tf::LookupException& ex) {
